+++ component/src/main/java/org/wso2/extension/siddhi/io/cdc/source/listening/InMemoryOffsetBackingStore.java
@@ -16,12 +16,13 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.io.cdc.source;
+package org.wso2.extension.siddhi.io.cdc.source.listening;
 
 import org.apache.kafka.connect.runtime.WorkerConfig;
 import org.apache.kafka.connect.storage.MemoryOffsetBackingStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.wso2.extension.siddhi.io.cdc.source.CDCSource;
 import org.wso2.extension.siddhi.io.cdc.util.CDCSourceConstants;
 
 import java.nio.ByteBuffer;
