+++ component/src/main/java/org/wso2/extension/siddhi/io/cdc/util/CDCSourceUtil.java
@@ -18,32 +18,20 @@
 
 package org.wso2.extension.siddhi.io.cdc.util;
 
-import org.apache.kafka.connect.connector.ConnectRecord;
-import org.apache.kafka.connect.data.Field;
-import org.apache.kafka.connect.data.Struct;
-import org.apache.kafka.connect.errors.DataException;
-import org.apache.log4j.Logger;
-import org.wso2.extension.siddhi.io.cdc.source.InMemoryOffsetBackingStore;
-import org.wso2.extension.siddhi.io.cdc.source.WrongConfigurationException;
+import org.wso2.extension.siddhi.io.cdc.source.listening.InMemoryOffsetBackingStore;
+import org.wso2.extension.siddhi.io.cdc.source.listening.WrongConfigurationException;
 import org.wso2.siddhi.query.api.exception.SiddhiAppValidationException;
 
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
-import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Random;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 /**
- * This class contains Util methods for the cdc extension.
+ * This class contains Util methods for the CDCSource.
  */
 public class CDCSourceUtil {
-    private static final Logger log = Logger.getLogger(CDCSourceUtil.class);
     public static Map<String, Object> getConfigMap(String username, String password, String url, String tableName,
                                                    String historyFileDirectory, String siddhiAppName,
                                                    String siddhiStreamName, int serverID, String serverName,
@@ -149,81 +137,6 @@ public class CDCSourceUtil {
     }
 
     /**
-     * Create Hash map using the connect record and operation,
-     *
-     * @param connectRecord is the change data object which is received from debezium embedded engine.
-     * @param operation     is the change data event which is specified by the user.
-     **/
-
-    public static Map<String, Object> createMap(ConnectRecord connectRecord, String operation) {
-
-        //Map to return
-        Map<String, Object> detailsMap = new HashMap<>();
-
-        Struct record = (Struct) connectRecord.value();
-
-        //get the change data object's operation.
-        String op;
-
-        try {
-            op = (String) record.get("op");
-        } catch (NullPointerException | DataException ex) {
-            return detailsMap;
-        }
-
-        //match the change data's operation with user specifying operation and proceed.
-        if (operation.equalsIgnoreCase(CDCSourceConstants.INSERT) &&
-                op.equals(CDCSourceConstants.CONNECT_RECORD_INSERT_OPERATION)
-                || operation.equalsIgnoreCase(CDCSourceConstants.DELETE) &&
-                op.equals(CDCSourceConstants.CONNECT_RECORD_DELETE_OPERATION)
-                || operation.equalsIgnoreCase(CDCSourceConstants.UPDATE) &&
-                op.equals(CDCSourceConstants.CONNECT_RECORD_UPDATE_OPERATION)) {
-
-            Struct rawDetails;
-            List<Field> fields;
-            String fieldName;
-
-            switch (op) {
-                case CDCSourceConstants.CONNECT_RECORD_INSERT_OPERATION:
-                    //append row details after insert.
-                    rawDetails = (Struct) record.get(CDCSourceConstants.AFTER);
-                    fields = rawDetails.schema().fields();
-                    for (Field key : fields) {
-                        fieldName = key.name();
-                        detailsMap.put(fieldName, rawDetails.get(fieldName));
-                    }
-                    break;
-                case CDCSourceConstants.CONNECT_RECORD_DELETE_OPERATION:
-                    //append row details before delete.
-                    rawDetails = (Struct) record.get(CDCSourceConstants.BEFORE);
-                    fields = rawDetails.schema().fields();
-                    for (Field key : fields) {
-                        fieldName = key.name();
-                        detailsMap.put(CDCSourceConstants.BEFORE_PREFIX + fieldName, rawDetails.get(fieldName));
-                    }
-                    break;
-                case CDCSourceConstants.CONNECT_RECORD_UPDATE_OPERATION:
-                    //append row details before update.
-                    rawDetails = (Struct) record.get(CDCSourceConstants.BEFORE);
-                    fields = rawDetails.schema().fields();
-                    for (Field key : fields) {
-                        fieldName = key.name();
-                        detailsMap.put(CDCSourceConstants.BEFORE_PREFIX + fieldName, rawDetails.get(fieldName));
-                    }
-                    //append row details after update.
-                    rawDetails = (Struct) record.get(CDCSourceConstants.AFTER);
-                    fields = rawDetails.schema().fields();
-                    for (Field key : fields) {
-                        fieldName = key.name();
-                        detailsMap.put(fieldName, rawDetails.get(fieldName));
-                    }
-                    break;
-            }
-        }
-        return detailsMap;
-    }
-
-    /**
      * Get the WSO2 Stream Processor's local path from System Variables.
      * if carbon.home is not set, return the current project path. (for test cases and for use as a java library)
      */
@@ -234,82 +147,5 @@ public class CDCSourceUtil {
             path = System.getProperty(CDCSourceConstants.USER_DIRECTORY);
         }
         return path;
-    }
-
-    /**
-     * Utility method which can be used to check if a given string instance is null or empty.
-     *
-     * @param field the string instance to be checked.
-     * @return true if the field is null or empty.
-     */
-    public static boolean isEmpty(String field) {
-        return (field == null || field.trim().length() == 0);
-    }
-
-    /**
-     * Create Hash map using the connect record and operation,
-     *
-     * @param resultSet is the resultset obtained by the cdcpollar.
-     **/
-
-    public static Map<String, Object> createMap(ResultSet resultSet) throws SQLException {
-        Map<String, Object> detailsMap = new HashMap<>();
-        ResultSetMetaData metadata = resultSet.getMetaData();
-
-        if (resultSet.next()) {
-            for (int i = 1; i <= metadata.getColumnCount(); i++) {
-                String key = metadata.getColumnName(i);
-                String value = resultSet.getString(key);
-                detailsMap.put(key, value);
-            }
-        }
-
-        return detailsMap;
-    }
-
-    /**
-     * Method which can be used to clear up and ephemeral SQL connectivity artifacts.
-     *
-     * @param rs   {@link ResultSet} instance (can be null)
-     * @param stmt {@link Statement} instance (can be null)
-     * @param conn {@link Connection} instance (can be null)
-     */
-    public static void cleanupConnection(ResultSet rs, Statement stmt, Connection conn) {
-        if (rs != null) {
-            try {
-                rs.close();
-                if (log.isDebugEnabled()) {
-                    log.debug("Closed ResultSet");
-                }
-            } catch (SQLException e) {
-                if (log.isDebugEnabled()) {
-                    log.debug("Error in closing ResultSet: " + e.getMessage(), e);
-                }
-            }
-        }
-        if (stmt != null) {
-            try {
-                stmt.close();
-                if (log.isDebugEnabled()) {
-                    log.debug("Closed PreparedStatement");
-                }
-            } catch (SQLException e) {
-                if (log.isDebugEnabled()) {
-                    log.debug("Error in closing PreparedStatement: " + e.getMessage(), e);
-                }
-            }
-        }
-        if (conn != null) {
-            try {
-                conn.close();
-                if (log.isDebugEnabled()) {
-                    log.debug("Closed Connection");
-                }
-            } catch (SQLException e) {
-                if (log.isDebugEnabled()) {
-                    log.debug("Error in closing Connection: " + e.getMessage(), e);
-                }
-            }
-        }
     }
 }
